+++ Editor/Common/Enums/ImportStatus.cs
@@ -1,0 +1,9 @@
+﻿namespace Editor.Common.Enums
+{
+    public enum ImportStatus
+    {
+        IMPORTING,
+        SUCCEEDED,
+        FAILED
+    }
+}